# docs/swagger.yaml
openapi: 3.0.1
info:
  title: Weather Forecast API
  description: Weather API application that allows users to subscribe to weather updates
    for their city.
  version: 1.0.0
servers:
- url: http://weatherapi.app/api
- url: https://weatherapi.app/api
tags:
- name: weather
  description: Weather forecast operations
- name: subscription
  description: Subscription management operations
paths:
  /weather:
    get:
      tags:
      - weather
      summary: Get current weather for a city
      description: Returns the current weather forecast for the specified city using
        WeatherAPI.com.
      operationId: getWeather
      parameters:
      - name: city
        in: query
        description: City name for weather forecast
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation - current weather forecast returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  temperature:
                    type: number
                    description: Current temperature
                  humidity:
                    type: number
                    description: Current humidity percentage
                  description:
                    type: string
                    description: Weather description
        "400":
          description: Invalid request
          content: {}
        "404":
          description: City not found
          content: {}
  /subscribe:
    post:
      tags:
      - subscription
      summary: Subscribe to weather updates
      description: Subscribe an email to receive weather updates for a specific city
        with chosen frequency.
      operationId: subscribe
      requestBody:
        content:
          application/json:
            schema:
              required:
              - city
              - email
              - frequency
              type: object
              properties:
                email:
                  type: string
                  description: Email address to subscribe
                city:
                  type: string
                  description: City for weather updates
                frequency:
                  type: string
                  description: Frequency of updates (hourly or daily)
                  enum:
                  - hourly
                  - daily
          application/x-www-form-urlencoded:
            schema:
              required:
              - city
              - email
              - frequency
              type: object
              properties:
                email:
                  type: string
                  description: Email address to subscribe
                city:
                  type: string
                  description: City for weather updates
                frequency:
                  type: string
                  description: Frequency of updates (hourly or daily)
                  enum:
                  - hourly
                  - daily
        required: true
      responses:
        "200":
          description: Subscription successful. Confirmation email sent.
          content: {}
        "400":
          description: Invalid input
          content: {}
        "409":
          description: Email already subscribed
          content: {}
  /confirm/{token}:
    get:
      tags:
      - subscription
      summary: Confirm email subscription
      description: Confirms a subscription using the token sent in the confirmation
        email.
      operationId: confirmSubscription
      parameters:
      - name: token
        in: path
        description: Confirmation token
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Subscription confirmed successfully
          content: {}
        "400":
          description: Invalid token
          content: {}
        "404":
          description: Token not found
          content: {}
  /unsubscribe/{token}:
    get:
      tags:
      - subscription
      summary: Unsubscribe from weather updates
      description: Unsubscribes an email from weather updates using the token sent
        in emails.
      operationId: unsubscribe
      parameters:
      - name: token
        in: path
        description: Unsubscribe token
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Unsubscribed successfully
          content: {}
        "400":
          description: Invalid token
          content: {}
        "404":
          description: Token not found
          content: {}
components:
  schemas:
    Weather:
      type: object
      properties:
        temperature:
          type: number
          description: Current temperature
        humidity:
          type: number
          description: Current humidity percentage
        description:
          type: string
          description: Weather description
    Subscription:
      required:
      - city
      - email
      - frequency
      type: object
      properties:
        email:
          type: string
          description: Email address
        city:
          type: string
          description: City for weather updates
        frequency:
          type: string
          description: Frequency of updates
          enum:
          - hourly
          - daily
        confirmed:
          type: boolean
          description: Whether the subscription is confirmed
x-original-swagger-version: "2.0"